First stub of timer update done
After every game that's been won or lost, a count of the length
of the game (in seconds) is printed.

Started (and probably finished) highscore.[ch], to save and import
highscores. Added a few functions to the new files misc.[ch] as
utility functions related to highscore.*. Makefile updated to handle
these new files.

The interface has still to be upgraded to display high scores and
to take a name every time a score is to be saved.